better configuration and setup of server plus static content serving
