+++ package.json
@@ -32,6 +32,7 @@
     "react": "^0.13.3",
     "react-select": "^0.5.4",
     "express": "~4.13.1",
-    "mongoose": "~4.0.6"
+    "mongoose": "~4.0.6",
+    "body-parser": "~1.13.2"
   }
 }
