+++ server/config/mongoose.js
@@ -1,8 +1,8 @@
 "use strict";
 var mongoose = require('mongoose');
 
-module.exports = function() {
-  mongoose.connect("mongodb://justinrassier:Oeve6Ne5sG2p@ds053160.mongolab.com:53160/persondb");
+module.exports = function(config) {
+  mongoose.connect(config.db);
 
   var Person = mongoose.model('Person', {
     firstName: String,
