+++ src/index.html
@@ -5,7 +5,7 @@
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css">
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap-theme.min.css">
   <link rel="stylesheet" href="styles/site.css" />
-
+  <script src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
   <script src="../lib/routie.min.js"></script>
 </head>
